refactor: page 이동 로직 제거, step으로 전환
