+++ package.json
@@ -26,8 +26,7 @@
     "@vanilla-extract/sprinkles": "^1.6.1",
     "clsx": "^2.1.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.2"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@storybook/addon-essentials": "^7.6.17",
