+++ src/app.tsx
@@ -1,25 +1,33 @@
-import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom'
-import { CardListPage } from '@/domains/card-list/page'
-import { CardRegisterPage } from '@/domains/card-register/page'
+import { useFunnel } from '@/hooks/use-funnel'
+import { CardInputContextProvider } from '@/steps/card-register/contexts/card-input-context'
+import { CardInputFormStep } from '@/steps/card-register'
+import { CardListStep } from './steps/card-list'
+import { CardRegisterCompleteStep } from '@/steps/card-register-complete'
 
-const router = createBrowserRouter([
-  {
-    index: true,
-    element: <Navigate to="/card-list" />,
-  },
-  {
-    path: '/card-list',
-    element: <CardListPage />,
-  },
-  {
-    path: '/card-register',
-    element: <CardRegisterPage />,
-  },
-])
+const 카드_목록 = 'card-list'
+const 카드_입력_폼 = 'card-input-form'
+const 카드_등록_완료 = 'card-register-complete'
 
 function App() {
+  const [Funnel, step, setStep] = useFunnel(카드_입력_폼)
+
   return (
-    <RouterProvider router={router} />
+    <CardInputContextProvider>
+      <Funnel.Root step={step}>
+        <Funnel.Step name={카드_목록}>
+          <CardListStep onClickRegister={() => setStep(카드_입력_폼)} />
+        </Funnel.Step>
+        <Funnel.Step name={카드_입력_폼}>
+          <CardInputFormStep
+            onClickPrev={() => setStep(카드_목록)}
+            onSubmit={() => setStep(카드_등록_완료)}
+          />
+        </Funnel.Step>
+        <Funnel.Step name={카드_등록_완료}>
+          <CardRegisterCompleteStep onClickConfirm={() => setStep(카드_목록)} />
+        </Funnel.Step>
+      </Funnel.Root>
+    </CardInputContextProvider>
   )
 }
