+++ src/components/template/header/header.tsx
@@ -1,21 +1,20 @@
 import { Flex, Text } from '@/components'
 import { vars } from '@/styles'
 import { IconChevronLeft } from '@tabler/icons-react'
-import { Link } from 'react-router-dom'
 
 export interface HeaderProps {
-  prevNavPath?: string
   title: string
+  onClickPrev?: () => void
 }
 
-export const Header = ({ title, prevNavPath }: HeaderProps) => {
+export const Header = ({ title, onClickPrev }: HeaderProps) => {
   return (
     <Flex as="header" width="100%" paddingX="24px" paddingY="36px">
       <Flex alignItems="center" gap="12px" color="gray900">
-        {prevNavPath && (
-          <Link to={prevNavPath}>
+        {onClickPrev && (
+          <Flex as="button" onClick={onClickPrev}>
             <IconChevronLeft size={24} color={vars.color.gray900} />
-          </Link>
+          </Flex>
         )}
         <Text variant="body1" color="gray900">
           {title}
