+++ src/steps/card-register-complete/index.tsx
@@ -6,15 +6,7 @@ export interface CardRegisterCompleteStepProps {
 }
 
 export const CardRegisterCompleteStep = ({ onClickConfirm }: CardRegisterCompleteStepProps) => {
-  const {
-    cardInput: { cardName, cardExpDate, cardCode },
-    resetCardInput,
-  } = useCardInputContext()
-
-  const handleclickConfirm = () => {
-    resetCardInput()
-    onClickConfirm()
-  }
+  const { cardInput } = useCardInputContext()
 
   return (
     <Flex width="100%" height="100vh" direction="column">
@@ -22,18 +14,12 @@ export const CardRegisterCompleteStep = ({ onClickConfirm }: CardRegisterComplet
         <Text as="h1" variant="heading1" marginBottom="48px">
           카드등록이 완료되었습니다.
         </Text>
-        <Card
-          cardSize="lg"
-          cardCode={cardCode}
-          cardExpDate={cardExpDate}
-          cardName={cardName}
-          marginBottom="24px"
-        />
+        <Card cardSize="lg" {...cardInput} marginBottom="24px" />
         <UnderlineInput width="240px" placeholder="카드 별명" contentAlign="center" />
       </Flex>
 
       <Flex justifyContent="flex-end" paddingX="24px" paddingY="32px" marginTop="auto">
-        <Text as="button" variant="body1" color="aqua" onClick={handleclickConfirm}>
+        <Text as="button" variant="body1" color="aqua" onClick={onClickConfirm}>
           확인
         </Text>
       </Flex>
