+++ src/steps/card-register/contexts/card-input-context.tsx
@@ -2,18 +2,17 @@ import { PropsWithChildren, createContext, useContext } from 'react'
 import { useCardInputState } from '@/steps/card-register/hooks/use-card-input-state'
 
 export interface CardInputContextValue {
-  cardInput: ReturnType<typeof useCardInputState>['cardInput']
-  setCardInput: ReturnType<typeof useCardInputState>['setCardInput']
-  resetCardInput: ReturnType<typeof useCardInputState>['resetCardInput']
+  cardInput: ReturnType<typeof useCardInputState>[0]
+  setCardInput: ReturnType<typeof useCardInputState>[1]
 }
 
 export const CardInputContext = createContext<CardInputContextValue | null>(null)
 
 export const CardInputContextProvider = ({ children }: PropsWithChildren) => {
-  const { cardInput, setCardInput, resetCardInput } = useCardInputState()
+  const [cardInput, setCardInput] = useCardInputState()
 
   return (
-    <CardInputContext.Provider value={{ cardInput, setCardInput, resetCardInput }}>
+    <CardInputContext.Provider value={{ cardInput, setCardInput }}>
       {children}
     </CardInputContext.Provider>
   )
