+++ src/steps/card-register/hooks/use-card-input-state.tsx
@@ -8,24 +8,18 @@ export interface CardInputState {
   cardPin: string
 }
 
-const INITIAL_CARD_STATE: CardInputState = {
-  cardCode: '',
-  cardExpDate: '',
-  cardName: '',
-  cardCVC: '',
-  cardPin: '',
-}
-
 export const useCardInputState = () => {
-  const [cardInput, setCardInput] = useState<CardInputState>(INITIAL_CARD_STATE)
+  const [cardInput, setCardInput] = useState<CardInputState>({
+    cardCode: '',
+    cardExpDate: '',
+    cardName: '',
+    cardCVC: '',
+    cardPin: '',
+  })
 
   const handleChangeCardInput = (name: keyof CardInputState) => (value: string) => {
     setCardInput(prev => ({ ...prev, [name]: value }))
   }
 
-  const resetCardInput = () => {
-    setCardInput(INITIAL_CARD_STATE)
-  }
-
-  return { cardInput, setCardInput: handleChangeCardInput, resetCardInput } as const
+  return [cardInput, handleChangeCardInput] as const
 }
